zed-edit-proxy/pyproject.toml
[project]
name = "zed-edit-proxy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.0",
    "httpx>=0.25.0",
    "pydantic>=2.4.0",
    "pyyaml>=6.0",
    "uvicorn[standard]>=0.24.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.12.9",
]

[tool.ruff]
line-length = 120
exclude = [".git", ".venv"]

[tool.ruff.lint]
fixable = ["ALL"]
